Change uint32_t -> size_t in *Enumerate functions
Follows https://github.com/webgpu-native/webgpu-headers/pull/142

Bug: dawn:160

+++ src/dawn_wire/client/LimitsAndFeatures.cpp
@@ -32,7 +32,7 @@ namespace dawn_wire { namespace client {
         return mFeatures.count(feature) != 0;
     }
 
-    uint32_t LimitsAndFeatures::EnumerateFeatures(WGPUFeatureName* features) const {
+    size_t LimitsAndFeatures::EnumerateFeatures(WGPUFeatureName* features) const {
         if (features != nullptr) {
             for (WGPUFeatureName f : mFeatures) {
                 *features = f;
